Fix SF 4 & 5 compatibility
