+++ Tests/EventListener/ReplyToHttpResponseListenerTest.php
@@ -11,7 +11,6 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
-use Symfony\Component\HttpKernel\Kernel;
 
 class ReplyToHttpResponseListenerTest extends TestCase
 {
@@ -34,7 +33,7 @@ class ReplyToHttpResponseListenerTest extends TestCase
         $event = new ExceptionEvent(
             $this->createHttpKernelMock(),
             new Request,
-            HttpKernelInterface::MAIN_REQUEST,
+            $this->getRequestType(),
             $expectedException
         );
 
@@ -66,7 +65,7 @@ class ReplyToHttpResponseListenerTest extends TestCase
         $event = new ExceptionEvent(
             $this->createHttpKernelMock(),
             new Request,
-            HttpKernelInterface::MAIN_REQUEST,
+            $this->getRequestType(),
             $reply
         );
 
@@ -94,7 +93,7 @@ class ReplyToHttpResponseListenerTest extends TestCase
         $reply = new HttpRedirect('/foo/bar');
         $response = new Response('', 302);
 
-        $event = new ExceptionEvent($this->createHttpKernelMock(), new Request, HttpKernelInterface::MAIN_REQUEST, $reply);
+        $event = new ExceptionEvent($this->createHttpKernelMock(), new Request, $this->getRequestType(), $reply);
 
         $converterMock = $this->createReplyToSymfonyResponseConverterMock();
         $converterMock
@@ -127,5 +126,14 @@ class ReplyToHttpResponseListenerTest extends TestCase
     protected function createHttpKernelMock()
     {
         return $this->createMock(HttpKernelInterface::class);
+    }
+
+    private function getRequestType(): int
+    {
+        if (defined(HttpKernelInterface::class . '::MAIN_REQUEST')) {
+            return HttpKernelInterface::MAIN_REQUEST;
+        }
+
+        return HttpKernelInterface::MASTER_REQUEST;
     }
 }
