+++ Tests/Functional/Form/Type/CreditCardTypeTest.php
@@ -16,7 +16,7 @@ class CreditCardTypeTest extends WebTestCase
     {
         parent::setUp();
 
-        $this->formFactory = static::getContainer()->get('form.factory');
+        $this->formFactory = static::$container->get('form.factory');
     }
 
     /**
