+++ Tests/Functional/Form/Type/GatewayConfigTypeTest.php
@@ -16,7 +16,7 @@ class GatewayConfigTypeTest extends WebTestCase
     {
         parent::setUp();
 
-        $this->formFactory = static::getContainer()->get('form.factory');
+        $this->formFactory = static::$container->get('form.factory');
     }
 
     /**
