+++ Tests/Functional/app/config/config_sf4.yml
@@ -1,0 +1,4 @@
+framework:
+  session:
+    storage_id: 'session.storage.mock_file'
+  validation: { enable_annotations: false }
