+++ Tests/Functional/app/config/config_sf5.yml
@@ -1,0 +1,7 @@
+framework:
+    session:
+        handler_id: null
+        cookie_secure: auto
+        cookie_samesite: lax
+        storage_factory_id: session.storage.factory.mock_file
+    validation: { enable_annotations: false }
